Collapse one class implementated interface.
